+++ app/src/main/java/com/example/instagram/IgViewModel.kt
@@ -10,7 +10,7 @@ import com.example.instagram.common.util.Constants.USERNAME
 import com.example.instagram.common.util.Constants.USERS
 import com.example.instagram.coroutineExtensions.combine
 import com.example.instagram.coroutineExtensions.stateInDefault
-import com.example.instagram.entities.User
+import com.example.instagram.models.User
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import dagger.hilt.android.lifecycle.HiltViewModel
