+++ app/src/main/java/com/example/instagram/auth/login/LoginViewmodel.kt
@@ -7,7 +7,7 @@ import com.example.instagram.common.extensions.ViewEventSinkFlow
 import com.example.instagram.common.util.Constants.USERS
 import com.example.instagram.coroutineExtensions.combine
 import com.example.instagram.coroutineExtensions.stateInDefault
-import com.example.instagram.entities.User
+import com.example.instagram.models.User
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.storage.FirebaseStorage
