+++ app/src/main/java/com/example/instagram/models/User.kt
@@ -1,7 +1,10 @@
-package com.example.instagram.entities
+package com.example.instagram.models
 
+import android.os.Parcelable
 import androidx.compose.runtime.Immutable
+import kotlinx.android.parcel.Parcelize
 
+@Parcelize
 @Immutable
 data class User(
     val userId: String? = null,
@@ -12,7 +15,7 @@ data class User(
     var following: List<String>? = null,
     var followers: List<String>? = null,
     var posts: List<String>? = null,
-) {
+) : Parcelable {
     fun toMap() = mapOf(
         "userId" to userId,
         "name" to name,
