+++ app/src/main/java/com/example/instagram/my_posts/MyPostsViewState.kt
@@ -9,8 +9,8 @@ import androidx.compose.runtime.setValue
 import com.example.instagram.common.extensions.OneTimeEvent
 import com.example.instagram.common.extensions.ViewEventSink
 import com.example.instagram.core_data.Fakes
-import com.example.instagram.domain.ig_domain.PostData
-import com.example.instagram.entities.User
+import com.example.instagram.models.PostData
+import com.example.instagram.models.User
 
 @Immutable
 data class MyPostsViewState(
