+++ app/src/main/java/com/example/instagram/profile/ProfileViewModel.kt
@@ -12,8 +12,8 @@ import com.example.instagram.common.util.Constants.POSTS
 import com.example.instagram.common.util.Constants.USERS
 import com.example.instagram.coroutineExtensions.combine
 import com.example.instagram.coroutineExtensions.stateInDefault
-import com.example.instagram.domain.ig_domain.PostData
-import com.example.instagram.entities.User
+import com.example.instagram.models.PostData
+import com.example.instagram.models.User
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.FirebaseFirestore
