+++ app/src/main/java/com/example/instagram/profile/ProfileViewState.kt
@@ -10,7 +10,7 @@ import androidx.compose.runtime.setValue
 import com.example.instagram.common.extensions.OneTimeEvent
 import com.example.instagram.common.extensions.ViewEventSink
 import com.example.instagram.core_data.Fakes
-import com.example.instagram.entities.User
+import com.example.instagram.models.User
 
 @Immutable
 data class ProfileViewState(
