+++ app/src/main/java/com/example/instagram/single_post/SinglePostViewModel.kt
@@ -9,8 +9,8 @@ import com.example.instagram.common.util.Constants.POSTS
 import com.example.instagram.common.util.Constants.USERS
 import com.example.instagram.coroutineExtensions.combine
 import com.example.instagram.coroutineExtensions.stateInDefault
-import com.example.instagram.domain.ig_domain.PostData
-import com.example.instagram.entities.User
+import com.example.instagram.models.PostData
+import com.example.instagram.models.User
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.QuerySnapshot
